added all routes accross all operations
